+++ src/lidia/aircraft.py
@@ -1,7 +1,9 @@
 import json
+from math import cos, sin, sqrt
 from pydantic import BaseModel
-from typing import Any, Dict, Optional
+from typing import Any, Dict, Iterable, List, Optional
 
+from .config import Config as LidiaConfig
 from .mytypes import VectorModel, IntFlagModel
 
 
@@ -105,9 +107,137 @@ class AircraftState(BaseModel):
         # HACK: JSON roundtrip is required, because there is no encoder configuration for .dict()
         return json.loads(self.json(models_as_dict=False, exclude={f for f in self.__fields__ if getattr(self, f) is None}))
 
+    def _matmul(matrix: Iterable[float], multiplicand: Iterable[float]) -> List[float]:
+        """Multiply 3x3 matrix by 3-element vector or 3x3 matrix"""
+        assert len(matrix) == 9
+        assert len(multiplicand) == 3 or len(multiplicand) == 9
+        if len(multiplicand) == 3:
+            m = matrix
+            x, y, z = multiplicand
+            return [
+                x * m[0] + y * m[1] + z * m[2],
+                x * m[3] + y * m[4] + z * m[5],
+                x * m[6] + y * m[7] + z * m[8],
+            ]
+        if len(multiplicand) == 9:
+            m = matrix
+            n = multiplicand
+            # aut
+            return [
+                m[0] * n[0] + m[1] * n[3] + m[2] * n[6], m[0] * n[1] + m[1] * n[4] + m[2] * n[7], m[0] * n[2] + m[1] * n[5] + m[2] * n[8],  # noqa
+                m[3] * n[0] + m[4] * n[3] + m[5] * n[6], m[3] * n[1] + m[4] * n[4] + m[5] * n[7], m[3] * n[2] + m[4] * n[5] + m[5] * n[8],  # noqa
+                m[6] * n[0] + m[7] * n[3] + m[8] * n[6], m[6] * n[1] + m[7] * n[4] + m[8] * n[7], m[6] * n[2] + m[7] * n[5] + m[8] * n[8],  # noqa
+            ]
+
+    def _transpose(matrix: Iterable[float]) -> List[float]:
+        """Transpose 3x3 matrix"""
+        assert len(matrix) == 9
+        m = matrix
+        return [
+            m[0], m[3], m[6],
+            m[1], m[4], m[7],
+            m[2], m[5], m[8],
+        ]
+
+    def _rotmat(att: Attitude) -> List[float]:
+        """Create 3x3 rotation matrix transforming body-frame vectors to outside frame"""
+        s, c = sin(att.roll), cos(att.roll)
+        rot_roll = [
+            1, 0, 0,
+            0, c, -s,
+            0, s, c,
+        ]
+        s, c = sin(att.pitch), cos(att.pitch)
+        rot_pitch = [
+            c, 0, s,
+            0, 1, 0,
+            -s, 0, c
+        ]
+        s, c = sin(att.yaw), cos(att.yaw)
+        rot_yaw = [
+            c, -s, 0,
+            s, c, 0,
+            0, 0, 1
+        ]
+
+        return AircraftState._matmul(rot_yaw, AircraftState._matmul(rot_pitch, rot_roll))
+
+    def xyz2ned(self, vec: XYZ) -> NED:
+        """Transform from body frame vector to outside frame coordinates
+
+        Assumes no rotation if `self.att` is missing"""
+        if self.att is None:
+            return NED.from_list([vec.x, vec.y, vec.z])
+        else:
+            return NED.from_list(AircraftState._matmul(AircraftState._rotmat(self.att), [vec.x, vec.y, vec.z]))
+
+    def ned2xyz(self, vec: NED) -> XYZ:
+        """Transform from outside frame to body frame coordinates
+
+        Assumes no rotation if `self.att` is missing"""
+        if self.att is None:
+            return XYZ.from_list([vec.north, vec.east, vec.down])
+        else:
+            return XYZ.from_list(AircraftState._matmul(AircraftState._transpose(AircraftState._rotmat(self.att)), [vec.north, vec.east, vec.down]))
+
+    def model_instruments(self, config: LidiaConfig):
+        """Generate values for instruments based on known parameters and configuration"""
+        self._model_ias(config)
+        self._model_gs(config)
+        self._model_alt(config)
+        self._model_ralt(config)
+
+    def _get_instr(self) -> Instruments:
+        if self.instr is None:
+            self.instr = Instruments()
+        return self.instr
+
+    def _model_ias(self, config: LidiaConfig) -> bool:
+        ias = None
+        if self.v_body is not None:
+            ias = self.v_body.x
+        elif self.v_ned is not None and self.att is not None:
+            v_body = self.ned2xyz(self.v_ned)
+            ias = v_body.x
+
+        if ias is not None:
+            self._get_instr().ias = ias * config.instruments.speed_multiplier
+            return True
+        return False
+
+    def _model_gs(self, config: LidiaConfig) -> bool:
+        gs = None
+        if self.v_ned is not None:
+            gs = sqrt(self.v_ned.north ** 2 + self.v_ned.east ** 2)
+        elif self.v_body is not None and self.att is not None:
+            v_ned = self.xyz2ned(self.v_body)
+            gs = sqrt(v_ned.north ** 2 + v_ned.east ** 2)
+
+        if gs is not None:
+            self._get_instr().gs = gs * config.instruments.speed_multiplier
+            return True
+        return False
+
+    def _model_alt(self, config: LidiaConfig) -> bool:
+        if self.ned is not None:
+            self._get_instr().alt = -self.ned.down * config.instruments.altitude_multiplier
+            return True
+        return False
+
+    def _model_ralt(self, config: LidiaConfig) -> bool:
+        if self.ned is not None:
+            if -self.ned.down <= config.instruments.radio_altimeter_activation:
+                self._get_instr().ralt = -self.ned.down * config.instruments.altitude_multiplier
+                return True
+        return False
+
 
 if __name__ == '__main__':
+    # run this from src/ folder like this: python -m lidia.aircraft
+    config = LidiaConfig()
     state = AircraftState()
     state.ned = NED.from_list([1, 2, 3])
     state.att = Attitude.from_list([4, 5, 6])
+    state.v_body = XYZ.from_list([5, 0, 0])
+    state.model_instruments(config)
     print(state.smol())